Help with tooltip for parameters / force display mode
KiwiHC16/Abeille/#138
showAllCommands: force display in widget
when set to yes is a mode where all commands & infos are shown in widget. (dev mode)
When set to no, normal display.(normal mode)

KiwiHC16/Abeille/#137
when the mouse hover the text, an help is displayed over it

Fix typo when removing the plugin.